Support for structured payee-based links in the memo field of transactions
This is done by adding a template for a link, e.g..
"https://www.heise.de/select/ct/archiv/%1" or "file:///.../%1" and a
regular expression for an identifier, e.g. "(\d{4}/\d{2})\b", at the
associated payee to capture the term "xxxx/yy". For this purpose, a new
register entry "Link Matching" was added to the payee.

In the memo field of the transaction associated with this payee, the
corresponding term, e.g. "2022/15", can be added to get a link marked
with "<...>" in the general ledger view.

The link can be opened, after selecting the corresponding transaction,
via the new entry "Open URL" in the context menu or via a keyboard
shortcut.

The current implementation supports capturing up to two independent patterns
e.g. the term "xxxx/yy" can be defined as "(\d{4})/(\d+)\b" to get links
at yearly subdivisions like: "file://<root>/bills/%1/vf/%2", where
%1 contains the year ('xxxx') and %2 contains the bill number ('yy').

For local file paths the url template also supports regular expressions
to find files of which the whole file name cannot be specified via the
determined identifier e.g. "file://<root>/bills/%1/vf/.*%2.*.pdf".

BUG:437481
FIXED-IN:5.2.0

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>